requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pypdf
sentence-transformers
faiss-cpu
numpy

